+++ extensions/easydict/src/releaseVersion/releaseNote.ts
@@ -2,37 +2,36 @@
  * @author: tisfeng
  * @createTime: 2022-07-03 22:10
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-10-13 12:23
+ * @lastEditTime: 2022-10-18 10:48
  * @fileName: releaseNote.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
 export const releaseNote = `
-## [v2.5.0] - 2022-10-12
+## [v2.6.0] - 2022-10-18
 
-本次更新主要支持了**现代汉语词典**，方便查询中文词语的拼音和释义了。
+本次更新主要支持了[扩展参数](https://developers.raycast.com/information/arguments)。
 
 ### ✨ 新功能
 
-- 新增支持有道现代汉语词典。
+- 新增支持了[Arguments](https://developers.raycast.com/information/arguments)。
+- 新增支持了乌克兰等 25 种语言，总计支持 48 种语言。
 
-![yi-1665582552](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/yi-1665582552.png)
+![arguments-1666060638](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/arguments-1666060638.png)
 
-![ai-1665582591](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/ai-1665582591.png)
+![hello-1666060655](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/hello-1666060655.png)
 
-![ai-1665582690](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/ai-1665582690.png)
-
+![Ukrainian-1666061433](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/Ukrainian-1666061433.png)
 
 ### 💎 优化
 
-- 优化了动作列表，现如果安装了 Eudic，会优先显示 'Open in Eudic' 动作，否则显示 'Copy Text' 动作。可以在设置中更改显示顺序。
-- 修改 Linguee 默认支持系统代理。
-- 修改所有列表都支持显示更多动作。
+- 优化了多行文本翻译时详情页面的显示效果。
+- 详情页面添加了 \`Copy Text\` 和打开网页查询的动作。
 
 ### 🐞 修复
 
-- 修复了许多已知问题。
+- 修复了一些已知问题。
 
 #### 如果觉得这个扩展还不错，给个 [Star](https://github.com/tisfeng/Raycast-Easydict) ⭐️ 支持一下吧 (^-^)
 
@@ -40,15 +39,15 @@ export const releaseNote = `
 
 ✨ New
 
-- Added support for Youdao Modern Chinese Dictionary.
+- Added support for [Arguments](https://developers.raycast.com/information/arguments).
+- Added support for 25 languages, including Ukrainian, total 48 languages.
 
-💎  Improvement
+### 💎 Improvement
 
-- Improved the action list, now if Eudic is installed, the 'Open in Eudic' action will be shown first, otherwise the 'Copy Text' action will be shown. You can change the display order in the settings.
-- Changed Linguee to support system proxy by default.
-- Changed all lists to support 'Show More Details'.
+- Improved the display effect of the details page when translating multiple lines of text.
+- Added \`Copy Text\` and open the web page query action to the details page.
 
-🐞 Fixes
+### 🐞 Fixes
 
-- Fixed many known issues.
+- Fixed some known issues.
 `;