+++ extensions/easydict/CHANGELOG.md
@@ -1,5 +1,27 @@
 # `Easydict` Changelog
 
+## [v2.6.0] - 2022-10-18
+
+✨ New
+
+- Added support for [Arguments](https://developers.raycast.com/information/arguments).
+- Added support for 25 languages, including Ukrainian, total 48 languages.
+
+![arguments-1666060638](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/arguments-1666060638.png)
+
+![hello-1666060655](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/hello-1666060655.png)
+
+![Ukrainian-1666061433](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/Ukrainian-1666061433.png)
+
+### 💎 Improvement
+
+- Improved the display effect of the details page when translating multiple lines of text.
+- Added \`Copy Text\` and open the web page query action to the details page.
+
+### 🐞 Fixes
+
+- Fixed some known issues.
+
 ## [v2.5.0] - 2022-10-12
 
 ✨ New